requirements.txt
python-dotenv==0.21.0
pynput==1.7.6
pyaudio==0.2.13
playsound==1.2.2
requests==2.31.0
pyautogui==0.9.53
PyGObject